V4.1

+++ Sava_Utils/edit_panel.py
@@ -60,6 +60,7 @@ def play_audio(idx, subtitle_list):
 
 def getworklist():
     try:
+        assert not Sava_Utils.config.server_mode
         c=os.listdir(os.path.join(current_path,"SAVAdata","temp","work"))
         return gr.update(choices=c,value=c[-1])
     except:
@@ -197,7 +198,9 @@ def apply_spk(speaker, page, subtitles: Subtitles, *args):
 
 def del_spk(name):
     try:
-        assert name not in ["", "None", [],None],"不得为空"
+        if Sava_Utils.config.server_mode:
+            raise RuntimeError("当前功能被禁止")
+        assert name not in ["", "None", [],None],"不得为空"        
         os.remove(os.path.join(current_path, "SAVAdata", "speakers",name))
         gr.Info(f"删除：{name}")
     except Exception as e:
